requirements.txt
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
snowflake-connector-python[pandas]==3.4.0
snowflake-sqlalchemy==1.5.1
sqlalchemy==1.4.49

uh65063-docker-spcs.registry.snowflakecomputing.com/orca_tracker/weather/orca_tracker_image_repo